# app/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Manage member and rewards",
    "version": "1.0.0",
    "title": "Member Reward API",
    "license": {
      "name": "MIT",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:80",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "memberReward",
      "description": "Manage Member and rewards\n"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/app/ping": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "pingServer",
        "summary": "Check server status.",
        "description": "Simple health check.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ping"
            }
          },
          "500": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Ping"
      }
    },
    "/members": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "createMember",
        "summary": "submit member details",
        "description": "Submitting member details in the system\n",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "A JSON object instance.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/memberRequestModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/memberResponseModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    },
    "/members/{memberId}": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "retrieveMember",
        "summary": "retrieve member details",
        "description": "Retrieve member details using member id\n",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "description": "memberId to retrieve member details",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/memberResponseModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      },
      "delete": {
        "tags": [
          "System"
        ],
        "operationId": "deleteMember",
        "summary": "delete member details",
        "description": "Delete member details using member id\n",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "description": "memberId to delete member details",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    },
    "/rewards": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "createReward",
        "summary": "submit new reward",
        "description": "Submitting new Reward details\n",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "A JSON object instance.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/rewardRequestModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/rewardResponseModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    },
    "/rewards/{rewardId}": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "retrieveReward",
        "summary": "retrieve reward details",
        "description": "Retrieve reward details using reward id\n",
        "parameters": [
          {
            "in": "path",
            "name": "rewardId",
            "description": "rewardId to retrieve reward details",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/rewardResponseModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      },
      "delete": {
        "tags": [
          "System"
        ],
        "operationId": "deleteReward",
        "summary": "delete reward details",
        "description": "Delete reward details using reward id\n",
        "parameters": [
          {
            "in": "path",
            "name": "rewardId",
            "description": "rewardId to delete reward details",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    },
    "/members/{memberId}/rewards/{rewardId}": {
      "put": {
        "tags": [
          "System"
        ],
        "operationId": "associateReward",
        "summary": "associate reward with member",
        "description": "Associate reward with member based on path parameters merberId and rewardId\n",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "description": "memberId to be associtaed with reward",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "rewardId",
            "description": "rewardId to be associated with member",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    },
    "/members/{memberId}/rewards": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "retrieveMemberRewards",
        "summary": "retrieve all rewards associated with a member",
        "description": "Retrieve all rewards associated with a member\n",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "description": "memberId to be associtaed with reward",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/rewardsResponseModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/errors"
            }
          }
        },
        "x-swagger-router-controller": "Apps"
      }
    }
  },
  "definitions": {
    "errors": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/errorModel"
          }
        }
      }
    },
    "errorModel": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "nessage to the consumer reagrding the error"
        }
      }
    },
    "ping": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "memberRequestModel": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/memberCommon"
        }
      }
    },
    "memberCommon": {
      "type": "object",
      "required": [
        "name",
        "email"
      ],
      "properties": {
        "name": {
          "description": "Full name of the member",
          "type": "string"
        },
        "email": {
          "description": "Email address of the member",
          "type": "string"
        },
        "contactNumber": {
          "description": "Contact number - mobile or landline",
          "type": "string"
        }
      }
    },
    "memberResponseModel": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/memberMaster"
        }
      }
    },
    "memberMaster": {
      "allOf": [
        {
          "$ref": "#/definitions/memberCommon"
        },
        {
          "type": "object",
          "properties": {
            "id": {
              "description": "Id of the member",
              "type": "integer"
            },
            "createdDate": {
              "description": "Created date of the member",
              "type": "string"
            },
            "updatedDate": {
              "description": "Updated date of the member",
              "type": "string"
            },
            "createdBy": {
              "description": "Created by of the record",
              "type": "string"
            },
            "updatedBy": {
              "description": "Updated by of the record",
              "type": "string"
            }
          }
        }
      ]
    },
    "rewardRequestModel": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/rewardCommon"
        }
      }
    },
    "rewardCommon": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "description": "Name of the reward",
          "type": "string"
        }
      }
    },
    "rewardResponseModel": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/rewardMaster"
        }
      }
    },
    "rewardsResponseModel": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/rewardMaster"
          }
        }
      }
    },
    "rewardMaster": {
      "allOf": [
        {
          "$ref": "#/definitions/rewardCommon"
        },
        {
          "type": "object",
          "properties": {
            "id": {
              "description": "Id of the reward",
              "type": "integer"
            },
            "createdDate": {
              "description": "Created date of the reward",
              "type": "string"
            },
            "updatedDate": {
              "description": "Updated date of the reward",
              "type": "string"
            },
            "createdBy": {
              "description": "Created by of the reward",
              "type": "string"
            },
            "updatedBy": {
              "description": "Updated by of the reward",
              "type": "string"
            }
          }
        }
      ]
    }
  }
}